+++ client/src/components/Forms/AdminCatAndSubca/UpadateSubCat.jsx
@@ -11,19 +11,36 @@ import {
 } from "@mui/material";
 import React from "react";
 import { useDispatch } from "react-redux";
+import Swal from "sweetalert2";
 import { deleteSubCategory, getSubCategories } from "../../../redux/actions";
+import EditIcon from "@mui/icons-material/Edit";
+import DeleteIcon from "@mui/icons-material/Delete";
 
 export const UpadateSubCat = ({ allSubCategories }) => {
   const dispatch = useDispatch();
 
   async function handleDeleteSubCat(e) {
     e.preventDefault();
-    await dispatch(deleteSubCategory(e.target.value));
+    Swal.fire({
+      title: "¿Está seguro de eliminar esta sub categoría?",
+      text: "Esta acción no se puede deshacer!",
+      icon: "warning",
+      background: "#DFDCD3",
+      showCancelButton: true,
+      confirmButtonColor: "#B6893E",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si, elimnar!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await dispatch(deleteSubCategory(e.target.value));
     await dispatch(getSubCategories());
+        Swal.fire("Eliminada!", "Esta sub categoría ha sido eliminada.", "success");
+      }
+    });
   }
 
   return (
-    <Box sx={{display:'flex', flexDirection:'column',alignItems:'center', m:3}}>
+    <Box sx={{display:'flex', flexDirection:'column',alignItems:'center', m:3, width:"100%"}}>
       <Typography variant="h5" sx={{m:3}}>Actualizar subcategorias</Typography>
       <TableContainer component={Paper} >
         <Table>
@@ -38,6 +55,8 @@ export const UpadateSubCat = ({ allSubCategories }) => {
               <TableCell>
                 <Typography>Categorias</Typography>
               </TableCell>
+              <TableCell>
+              </TableCell>
             </TableRow>
           </TableHead>
           {allSubCategories?.map((sc) => {
@@ -47,11 +66,15 @@ export const UpadateSubCat = ({ allSubCategories }) => {
                 <TableCell>{sc.description}</TableCell>
                 <TableCell>{sc.categories[0]?.name}</TableCell>
                 <TableCell>
+                  
                   <Button
                     value={sc.id}
                     onClick={(e) => handleDeleteSubCat(e)}
                     name="delete"
-                    //   startIcon={<DeleteIcon />}
+                    startIcon={<DeleteIcon />}
+                    variant="contained"
+                    color="ambar3"
+                    size="small"
                   >
                     Eliminar
                   </Button>
